libcamera2: [WORKAROUND] remove ZSL guard for 8MP case
BZ: 104745

The camera firmware bug in CSS 20130420 cannot be hit with 8MP
picture-size, so do not apply the ZSL guard in this case. This
allows to avoid both the 6MP continuous shooting regression as
well as shot2shot PnP regression with 8MP picture size.

Note: this should NOT be ported to mainline as this is highly
CTP/CSS1.5 specific and applies to R42-stable only.

Change-Id: Ibf35498eda5d655b344d1ea910d9271db19ba1bb
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/105732
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Vtminen, LassiX <lassix.vtminen@intel.com>
Reviewed-by: Mandy, Arnaud <arnaud.mandy@intel.com>
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>