libcamera2: fixes saving/restoring flash modes
BZ: 104222

The flash icon in the UI was kept hidden due to not correctly restoring
the flash values that were saved when exiting from previous mode.

The flash modes were not correctly restored before due to
Social Camera bug, which caused the camera app to send us old
parameter values, that had the old value of "flash-mode-values"
string. That caused our flash restoring logic to be a bit lacking
and caused the issue with the fixed Social Camera 2.2 app.

Change-Id: Iaac64fbe23a3e3487132ecf52c4bcf82e3ca5756
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>
Reviewed-on: http://android.intel.com:8080/106503
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>