[PORT FROM MAIN] libcamera2: reduce observer sleep to recover faster from starving
BZ: 106200

This patch reduces the observer sleep to 33ms to recover faster
from starving. The actual time it waits in total for buffers to
be queued back to ISP, is not changed, since it is compensated by
increasing the retry count.

What was happening before this change, was that during the
original 200ms sleep several frames were queued back to ISP and
filled, and after the long sleep, the observer dequeued those
ready frames in a burst, and the starving and sleep then happened
again.

With a finer grained, shorter sleep, the queuing of buffers back
to ISP has still enough time to run, and the ISP will not fill
several frames while the observer sleeps. Recovering from the
initial poll failure is thus rather immediate.

Change-Id: I96312b3efe40ba15d4ff8d4b385648e354a833bf
Orig-Change-Id: I85690e4e6d53c217b7f60ceb6318a52631044610
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Reviewed-on: http://android.intel.com:8080/107249
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Vtminen, LassiX <lassix.vtminen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>