libcamera2: SmartShutter related S2S enhancement
BZ: 106681

Problem: isSmartRunning() is called during every takePicture.
Sometimes, but not always, PostProcThread was handling some
other message when the isSmartRunning call happened, and
responding to the message took a few milliseconds.

This patch makes PostProcThread::isSmartRunning() to run in
the context of the ControlThread, returning immediately, thus
making the call that is run during takePicture to happen faster.

This is made possible by making the starting and stopping of
smart shutter synchronous. Since ControlThread is the only one
changing the start/stop -state of smart shutter, now that it is
done synchronously the queries for isSmartRunning can be done
in the caller context by returning the related state variable
directly.

Change-Id: I84beddaa295758eca13883b62e963a638c08ef10
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Reviewed-on: http://android.intel.com:8080/106807
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>