libcamera2: move  adjust-fps frame skiping to AtomISP
BZ: 107252

Problem: Recent change has broken timelapse functionality.
The frame skipping logic to adjust from sensor fps to user
requested fps was in PreviewThread.
There was an assumption that preview thread gets first the
preview frame and it then decides whether to skip or not.

This assumption was broken by a recent patch that changed the
order how ControlThread and PreviewThread are attached to
the PreviewStreamSource

Solution:
Move the skipping logic to the PreviewStreamSource.
Remove the fps logic from PreviewThread.
In this way the fps for video and preview is ensured to be
the same regardless of what is the attachment order
to the PreviewStreamSource observer.

Change-Id: I88c81fc84923719a0fa53d962a1828b1364b2dc3
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/107175
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>