libcamera2: return to synchronous snapshot allocation
BZ: 105799

Problem:
Having the snapshot allocation asynchronous is not worth the
side effects. With the current model we are open to situations
where buffers are re-allocated while being used.
This translates in random segfaults in places that use snapshot
buffers.

There is no big gain in making this asyncrhonous. The new
snapshot buffer flow is such that we make less allocations
and less queries to PictureThread, because we know in ControlThread
how many we allocated.

Solution:
Make snapshot buffer allocation syncrhonous.
This reduces the complexity and the potential race conditions
where allocations triggered before free buffers currently in use.

Care needs to be taken now in ControlThread where the new method
is called. It has to be in a place where it doesn't impact
KPI metrics

Change-Id: I1702dacc7382b09955c67c994029076c7b173a98
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/107182
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>