[PORT FROM R42] libcamera2: always use correct counter to determine preview starving
BZ: 102368

PreviewStreamSource was using AtomISP::dataAvailable() for determining
when preview flow is starving from buffers. In video mode, this function
returns the number of buffers queued in video-pipe. While video pipe uses
the same amount of buffers, the flow is controlled asynchronously by
ControlThread. Effectively this made the logic to work inconsistently and
open for conditions where ISP starves and timeouts in occasions where it
can be prevented by sleeping.

Change-Id: Idb46316460352b8f95119213dae20e1aee5c524c
Orig-Change-Id: I6b32b89b05f01cd865d308ee49a3cbcde95ea0e5
Signed-off-by: Teemu Tuominen <teemux.tuominen@intel.com>
Reviewed-on: http://android.intel.com:8080/107533
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>