libcamera2: clean-up: Use AtomBuffer.dataPtr consistently
BZ: 106338

Since the introduction of the AtomBuffer field dataPtr
the usage of the buff->data is not encouraged. Since dataPtr
abstracts where the actual data is located, independently
of whether the buffer is shared or not.

Still there were some places where this was used.
This patch cleans this up.

Change-Id: I0c52c66734056f48494ecfc14237926dcb880219
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/106572
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>