[PORT FROM R42] libcamera2: fix mem in libva due to bad usage
BZ: 106142

Problem: in HDR case we create an ad-hoc liVA context to encode
the output from the HDR processing.
In this scenario the single VASurfaceId used for this new
context was not stored correctly in the array used during surface
destruction
This produced that the surface destruction to  fail when restoring original
context. The failed surface destruction created a memory leak in kernel

Solution:
store the surface Id in the correct array

Change-Id: Idf8531d5aade836b98a75680a07f3f060518d2a9
Orig-Change-Id: Ie5e88c24e3cbae8aecf882d3bb06871b048ceb75
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/107531
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>