libcamera2:[ARCH] use new 3A stats event to drive 3A thread
BZ: 110848

[ARCH]= Architectural significant change

Until now we have been using the preview frame dequeue
to run 3A thread. with the move to Off-line processing mode
the preview frames come later and therefore the 3A processing
loop is longer.

A recent optimization in FW allows to receive an event when
the 3A statistics are ready, ahead of the preview frame.

We use now this event to trigger 3A processing to achieve
shorted loop and improve the 3A responsiveness.

Change-Id: I42d3bcd085c80b2993da927328031ae6f10b4b64
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/109176
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Niemimuukko, Ukri <ukri.niemimuukko@intel.com>
Reviewed-by: Vtminen, LassiX <lassix.vtminen@intel.com>
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>