libcamera2:Fix pre-flash sequence with new 3AThread event
BZ: 110848

Since now we drive the 3A loop from the 3A statistics ready event
some methods have misleading names. This patch renames the old
newFrame and handleMessageNewFrame to newStats

Also  the pre-flash sequence was the only part that needed to be driven from
preview frames rather than 3A stats. With the previous change
this was broken.
To fix it we subscribe the 3AThread to preview frame notifications
only when we enter the pre-flash sequence.
The new newFrame methods in 3AThread are only used during the
pre-flash sequence now.

Change-Id: I673867f2d66c40f6c8b567b0f708087fbd307fc1
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/109177
Reviewed-by: Vtminen, LassiX <lassix.vtminen@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>