libcamera2: 3AThread stores only last SOF time
BZ: 110848

With the recent changes to drive 3A loop from 3A statistics
ready event the driver is still unable to provide
matching sequence numbers for the SOF and 3A stats event
This means that we cannot be 100% sure that the SOF event
with sequence number N belongs to the same frame as the
3A stats ready event with sequence number N.

So the matching of sequence numbers we have in the HAL
is pointless until we can guarantee this. In order to
do so, changes in the CSS FW are required ETA unknown.
For this reason we revert to providing the latest SOF
time and we leave it to the 3A algorithm (AF mainly)
to figure it out when the SOF for the statistics it
currently has really happen

We fast-track the SOF event processing in 3A Thread
to avoid being serialized with 3A stats event.
For that reason we use a mutex to protect the SOF
 event time

When the support from  FW and fix in diver is available
we can provide those matching SOF and 3ASTATS events
to the 3A algorithms, but for the time being this will
do.

Depends on:

Change-Id: Ib93c9744fc63b192f520b8dac207edba1aeeb2bf
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/109178
Reviewed-by: Vtminen, LassiX <lassix.vtminen@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>