libcamera2: update snapshot pixel format after allocation
BZ: 88559

Pixel format for snapshot is currently hardcoded to NV12
This format is used between camera and JPEG encoder.

In cases where raw capture is enabled JPEG encoding is broken,
(JPEG's are green if raw bayer is enabled)
this is a known limitation of how ISP FW handles bayer capture.

In cases where RAW capture is enabled via parameter during runtime
the pixel format was not updated correctly.

Solution:
Update the pixel format in the AtomIsp class every time we
allocate the snaphsot buffers.

Change-Id: I9986785d58ae5cf4452e14dc3b05aacd57076b28
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/110235
Reviewed-by: Han, Qiang <qiang.han@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>