libcamera2: fix the continuous-mode picture size limitation
BZ: 111565

The current limitation (see PSI BZ 73112) for small picture
sizes and continuous mode, is not fully correct. The correct
limitation is that picture size must be at least the same as
preview size.

Also added proper TODO markers to make it clear this is a temporary
measure until we have a CSS version that can handle these cases.

Change-Id: Ia19da171a2d5a67d261668d60315920cfc8bc339
Signed-off-by: Kai Vehmanen <kai.vehmanen@intel.com>
Reviewed-on: http://android.intel.com:8080/110381
Reviewed-by: Niemimuukko, Ukri <ukri.niemimuukko@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>