libcamera2: GPU scaler
BZ: 109245

This patch adds a GPU Scaler to libcamera2. The scaler is wrapped
into a ScalerService class which has its own thread and message queue
in order to make it thread safe and accessible from all other camera
HAL threads easily. Note: buffers that are used in the GPU scaler
must be allocated from graphics and registered before use.

Change-Id: I917924b580c227803b409a4512210830683757bf
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Signed-off-by: Ted Painter <ted.painter@intel.com>
Signed-off-by: Tao T Wu <tao.t.wu@intel.com>
Reviewed-on: http://android.intel.com:8080/106398
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>