libcamera2: HAL ZSL
BZ: 109245

This patch adds the HAL ZSL feature to libcamera2. The idea in this
implementation is to configure the driver to the maximum resolution in
the driver with the same aspect ratio as is requested for
preview/snapshot. Then a number of buffers is stored and held in the
HAL, and only part of buffers is kept in the driver, while preview
runs. Preview frames are downscaled from the HAL ZSL buffers with
the GPU, and when buffers are being passed, the oldest buffer is
always put back to the ISP.

Currently, when snapshot capture is requested, the oldest image that
is currently held in HAL is selected and scaled with the GPU. Same is
done for the postview image.

depends-on: Ie6605035e88412846ed19189b1b498bb4a337d47 in
device/intel/clovertrail

Change-Id: I70efa94d0c19907d8044b4a068f1a5ddb28b54dd
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Reviewed-on: http://android.intel.com:8080/110693
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>