libcamera2: Add support for VictoriaBay NVM data
BZ: 113139

In order to distinguish SaltBay and VictoriaBay imx135
sensors in NVM parser, also the vendor-platform-product
string needs to be passed to the parser.

This patch also moves the vendor-platform-product
creation function from CameraConf to PlatformData to make
it available for AtomAIQ and any other class that might
need it.

Change-Id: Ia4cbaef099a3e066c3a7a645665442f75b61d2b4
Signed-off-by: Kaisa Taipale <kaisa.taipale@intel.com>
Reviewed-on: http://android.intel.com:8080/111419
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>