libcamera2: fix rounding error in JPEG scaled quantization matrix
BZ: 114276

The variables involved in the multiplication to calculate the
scaled quantization value are 8 and 16 bit wide. In some combination
of low Q factor this multiplication goes over 2^16. for that
reason we need to up-cast the multiplication before the division
This error only appears on really low Q factors (<5) so it is not
a noticeable problem
Nevertheless needs to be fixed

Change-Id: I7df62fb47c690c5de63a4cd91fb0a54a72c44dee
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/112146
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Vtminen, LassiX <lassix.vtminen@intel.com>
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>