libcamera2: cleanup AtomAIQ::runAeMain() with new delayed results logic
BZ: 44999

Introduces AtomFifo based replacement for maintaining history of AEC
results in AtomAIQ. Change adds helper methods for utilizing the AtomFifo
as a results history queue. This is done essentially to ease integration of
further feature additions to AtomAIQ AEC.

Change clarifies and documents the logic in runAeMain() specific to a problem
of synching sensor parameters with statistics. Userspace synching solution is
used when sensor metadata is not available or reliably mapped with statistics.

AEC needs input parameters containing its own results as a feedback
corresponding to what is activate in sensor for a certain frame wherefrom the
statistics are read. These feedback results need to be aligned with exposure-
and gain-applying latencies introduced by sensors, plus the overall latencies
with 3A processing.

This change retains the old functionality with fixed feedback-delay of 2
and denotes that further work may address to reliability and
configurability of this delay.

AtomFifo originally from abandoned change
http://android.intel.com:8080/94239

Change-Id: Ief24d9efa9481b5b955877ef944be0081e75086c
Signed-off-by: Teemu Tuominen <teemux.tuominen@intel.com>
Signed-off-by: Yoshio Wada <yoshio.wada@intel.com>
Reviewed-on: http://android.intel.com:8080/113808
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>