libcamera2: using AtomDvs as IAtomIspObserver
BZ: 112523

Enables AtomDvs to be used as an observer for the preview stream events,
allowing for better synchronization between the frames and DVS info.
Better synchronization means that we will not be facing missing DVS
statistics (so often). Also, helps to avoid the fetching DVS information
for corrupt frames, which used to be the case for 0th frame when video
preview was started.

Change-Id: I8328970d4969c01c34d2a0d881a06533244ccfe4
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>
Reviewed-on: http://android.intel.com:8080/113102
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>