libcamera2: add support for File injection to new device clases
BZ: 109307

This patch implements file injection support using the new V4L2VideoNode
class.
An object of this class is created and the implementation of the file
injection feature is changed to use external buffers (a.k.a. USER_PTR)
instead of MMAP buffers to simplify the implementation of the
VideoNode class.
Support for MMAP buffers is not implemented yet and not used in HAL
currently, hence the change

The video node created is the ony OUTPUT video node in the HAL

Change-Id: I81068a263be67da51cc029c187cf8c448e1e6ade
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/114743
Reviewed-by: Han, Qiang <qiang.han@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>