libcamera2: fixes 13MP burst preview updating
BZ: 118848

The preview for 13MP burst was not updating, due to the picture
size 4160x3104 not being exactly 4:3 size. Therefore the postview
image did not get rendered on the preview, and the preview did not update.
Taking into use the "aspect tolerance" for checking the difference between
snapshot and postview aspect ratios, to allow less strict postview rendering.

Change-Id: Iddb7619c4d7e13cd0648f6bf5008cd5ca1577241
Signed-off-by: Lassi Väätämöinen <lassix.vtminen@intel.com>
Reviewed-on: http://android.intel.com:8080/115389
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>