libcamera2: change the hw jpeg encoder wrapper to meet the new libva interface
BZ: 119593

Issue:
the libva's interface for JPEG encoder has been changed,
so the hw jpeg encoder wrapper code should be changed to meet the new libva interface.

Solution:
1. meet the new libva interface.
2. make the jpeg quality factor can be configurated for jpeg quality.
We can use the libva's jpeg quality factor. And we also can use jpeg quality factor which is in the JPEGHwEncoder, not in libva.

Depends on:
115725 115742

Change-Id: I66455e264a0f1fae15b7181205f2615d6ad9b36b
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/115878
Reviewed-by: Sun, Jing A <jing.a.sun@intel.com>
Reviewed-by: Chen, Lei A <lei.a.chen@intel.com>
Reviewed-by: Shi, PingX <pingx.shi@intel.com>
Tested-by: Shi, PingX <pingx.shi@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>