libcamera2: generalize graphic buffer allocation
BZ: 88559

This is ground work to allow the allocation of graphic buffers to
be used for other uses.
Right now the implementation was hardcoded assuming that it was
for preview buffers.

With this patch the allocation receives more information to fill
up the AtomBuffer accordingly

It also fixes the case where RAW capture is triggered. This scenario
was not taken into account and crashed the HAL.

Change-Id: Iad4206073c0e41aca44b4765a69e45b2bb080d4c
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/116417
Reviewed-by: Han, Qiang <qiang.han@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>