libcamera2: prevent continuous capture of corrupted image
BZ: 120282

We are capturing a corrupted (greenish) frame in the following
scenario:
- preview just started
- capture is requested

In this situation the frame we get for teh snapshot does not
wait for the first frames to skip.
This initial frame skips due to sensor initialization are not
accounted for when requesting the snapshot

This patch takes those skips into account when requesting a capture
by adding the number of frame-skips to  the offset.
The continuous capture offset is essentially a way to ignore the first
capture frames

Change-Id: Idb07df6d8fe1c437951fe29805b248153a069bab
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/116582
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>