libcamera2: Fix the failure of CTS CameraTest#testSceneMode and testFocusAreas
BZ: 124085

Root cause:
The front camera doesn't support the focus mode now, but we set focus modes to it.
testSceneMode: When the cts tests the focus modes in different scene modes, it fails.
testFocusAreas: When the cts tests the focus areas, it fails.

Solution:
Add an evaluation funcion isFixedFocusCamera to decide whether it focus mode is "fixed".

Dependency: 117122

Change-Id: Id9ae3ba7c2cd313a92c1e258a3c88ee337a033fb
Signed-off-by: Yu CUI <yux.cui@intel.com>
Reviewed-on: http://android.intel.com:8080/117628
Reviewed-by: Patale, Pavan <pavan.patale@intel.com>
Tested-by: Patale, Pavan <pavan.patale@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>