libcamera2: fix panorama validation when extras are not enabled
BZ: 116405

Problem: panorama live preview size validation was failing when
the intel extras flag was set to false in the board config makefile.
The failure of the validation was preventing other
settings from being set by the application succesfully.

Solution: stop putting the panorama live preview size to the
parameter string in HAL, so that the setting coming back from the
application is not validated against an empty supported live
preview list.

Change-Id: I7b4dbb5f91343129ebe73fec429e4666a80933ad
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Reviewed-on: http://android.intel.com:8080/122434
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Reviewed-by: Kaikumaa, Timo <timo.kaikumaa@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>