libcamera2: [REVERT ME] Hardcoding flip setting for imx132
BZ: 126620

This patch is for flip setting of imx132 on Saltbay PR2. Because PR1(with
OV9724) and PR2(with imx132) share the same camera_profiles.xml, thus using
this hardcoded way for imx132.

This patch should be reverted when PR1 phased out and this setting moved to
camera_profiles.xml.

Change-Id: Id445f247e3cf3271120df4517d66fb276023d7ae
Signed-off-by: Bin Han <bin.b.han@intel.com>
Reviewed-on: http://android.intel.com:8080/122923
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Niemimuukko, Ukri <ukri.niemimuukko@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>