libcamera2: make getRecordingFrame non-blocking
BZ: 128501

Problem:
After ISP timeout the driver puts its buffers into its outputqueues and
the number of buffers for preview and recording does not usually match.
HAL side can only dequeue as many buffers as are available, without
blocking the ControlThread. After ISP timeout the ControlThread started
to block in getRecordingFrame with all kinds of message passing
related symptoms that manifested in the end as a lower video framerate.

Solution:
This patch modifies the getRecordingFrame API so that it is
non-blocking. ControlThread can then recover from ISP timeouts if
getRecordingFrame returns without a frame. Thanks to the
observer notifications, in normal cases there will always be frames,
and only after ISP timeout will the driver be dequeued dry of recording
buffers. With this patch in place, the unnecessary dequeuing is avoided
after the ISP timeout, and ControlThread will not be blocked.

Change-Id: I52081dd074915b06cfc6e31f26cd5a3b9d5aa917
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Reviewed-on: http://android.intel.com:8080/124412
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Lampila, KalleX <kallex.lampila@intel.com>
Reviewed-by: Govindapillai, Vinod <vinod.govindapillai@intel.com>
Reviewed-by: Kaikumaa, Timo <timo.kaikumaa@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>