libcamera2: separate postview allocation for non-hal-zsl
BZ: 129532

This patch separates the postview allocation for the non-hal-zsl
use-case so that all the memory is simply allocated via callbacks, and
the gfx allocation is not done.

The HAL ZSL case is fixed so that the actual image data is delivered to
the application instead of just the dummy 1 byte.

Change-Id: I95cc9018d5bc5f9046fee7864f5edc31e09ee496
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Reviewed-on: http://android.intel.com:8080/124665
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Kruger, Jozef <jozef.kruger@intel.com>
Tested-by: Kruger, Jozef <jozef.kruger@intel.com>
Reviewed-by: Han, Qiang <qiang.han@intel.com>
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>