[PORT FROM MAIN]libcamera2: separate postview allocation for non-hal-zsl
BZ: 129532

This patch separates the postview allocation for the non-hal-zsl
use-case so that all the memory is simply allocated via callbacks, and
the gfx allocation is not done.

The HAL ZSL case is fixed so that the actual image data is delivered to
the application instead of just the dummy 1 byte.

Change-Id: Ide5d46afcd4f825c6d58d2223aa4a4c54af06bc4
orig-Change-Id: I95cc9018d5bc5f9046fee7864f5edc31e09ee496
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Signed-off-by: Yu CUI <yux.cui@intel.com>