[PORT FROM Main] libcamera2: fix workaround restoring the previous flash results
BZ: 129758

Issue:
Initial exposure recalculation in mode switch resets the flash part
of AE results. The previous flash results are restored to feed the
correct initial values for flash to be used with the main snapshot.
This restoring was done with pointer re-assigning that leads to
undefined use of memcpy.

Solution:
Copy the struct instead.

Change-Id: Iec3b4738ddc5ce4b7a64098267d9ad16e710c7bb
Orig-Change-Id: I945f12cad11f1abdbb87afb357d3313d4e7c2a95
Signed-off-by: Teemu Tuominen <teemux.tuominen@intel.com>
Reviewed-on: http://android.intel.com:8080/136755
Reviewed-by: Hu, Gang A <gang.a.hu@intel.com>
Tested-by: Hu, Gang A <gang.a.hu@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>