libcamera2: fix workaround restoring the previous flash results
BZ: 129758

Issue:
Initial exposure recalculation in mode switch resets the flash part
of AE results. The previous flash results are restored to feed the
correct initial values for flash to be used with the main snapshot.
This restoring was done with pointer re-assigning that leads to
undefined use of memcpy.

Solution:
Copy the struct instead.

Change-Id: I945f12cad11f1abdbb87afb357d3313d4e7c2a95
Signed-off-by: Teemu Tuominen <teemux.tuominen@intel.com>
Reviewed-on: http://android.intel.com:8080/125563
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Niemimuukko, Ukri <ukri.niemimuukko@intel.com>
Reviewed-by: Han, Qiang <qiang.han@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>