libcamera2: use previous flash decisions for main flash
BZ: 126542

This patch makes the main flash to be used if the auto focus assist
light was on just before the capture, or if the main flash was
fired just before this capture.

The time the assist light start decides the flash usage, is set to
2000ms.
The time the previous flash decides the flash usage, is set to 1300ms.

Change-Id: Iae55cd2e6ee04b7548676dd9b11170ad14f860fe
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Reviewed-on: http://android.intel.com:8080/125738
Reviewed-by: Verdu, Miguel <miguel.verdu@intel.com>
Reviewed-by: Vtminen, LassiX <lassix.vtminen@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>