libcamera2: fix capture sub state machine
BZ: 126905

ULL introduced proprietary capture sequence where preview re-start is
allowed before application receives the final jpeg. A specific checking
of UltraLowLigth::isProcessing() does not cover the problematic, since
composition will follow the encoding phase. CaptureSubState is common
key to decide when to cancel and cleanup

In this patch we force ongoing post-capture processing items to complete
when there is a stop preview or when preview is re-started.
This ensures that all snapshot buffers are returned when there is
a change of parameters that triggers a preview re-start (like when we go to
HDR) or during a normal shutdown sequence.

The capture state machine is now consistently transitioning to IDLE in
all shooting modes when the captures are completed.
The order between:
 - encoding done
 - picture-done
remains dependent on the shooting mode.

Change-Id: Idc1541df0c382b327b6fa69504463ac68c76b299
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/126472
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Niemimuukko, Ukri <ukri.niemimuukko@intel.com>
Reviewed-by: Vtminen, LassiX <lassix.vtminen@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>