libcamera2: Add general way to query post-capture-processing status
BZ: 126905

This was currently done  by quering the ULL class.
Now we generalize for any ProcessingItem in the thread class.
The query is protected by mutex, so it is thread safe.

This patch also generalizes the cancel method so that it
does not require the instance of the ProcessingItem. It
stores it  internally.

Change-Id: I31bbad530816d344a91c43cd9a0cdc729de7db42
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/126471
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Vtminen, LassiX <lassix.vtminen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>