libcamera2: ULL: reset state when input buffers are retrieved
BZ: 126905

In case when ULL capture is cancelled the ULL status is not reset to
a valid value. This prevent ULL to be fired again after such a cancelation

This patch transitions the ULL algo to INIT after the input buffers used for
one ULL iteration are retrieved. It is assumed the algo is ready for a new
capture with similar configuration.

It also resets the frame status of those input frames to OK so that they are
recycled correctly in the control thread

Change-Id: Ibb3ce8679dc0cddab2424284e6998c549d6495e9
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/126491
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Vtminen, LassiX <lassix.vtminen@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>