[PORT FROM R42B] libcamera2: change the raw buffer number for the continuous mode
BZ: 131552

Issue:
the continuous capture will hang in the google camera application.
because the fw has been changed for performance reasons, so the old value 4 to the css2.0 is not enough.
we need to a bigger value for css2.0
But we need to keep the old value for the css1.5

Solution:
change the raw buffer number for the continutous mode in the camera HAL

Depends on:
None

Orig-Change-Id: I2e9511692a854eb77b9b1155e19cef6fa3f90376
Change-Id: I2e9511692a854eb77b9b1155e19cef6fa3f90376
Signed-off-by: Yang Liang <liang.l.yang@intel.com>
Reviewed-on: http://android.intel.com:8080/126553
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Zheng, Hongjiang A <hongjiang.a.zheng@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>