[PORT FROM MAIN] libcamera2: Ensure consistent RAW format selection during RAW capture
BZ: 130090

RAW Bayer format is not the same for all platforms. Selection done in the HAL
was assuming it is. It needs to be queried from the sensor.

This patch makes used of the newly added method in V4L2VideoNode class
to query the supported sensor formats.
AtomISP stores the formats and then it selects the first Bayer format
as prefered Raw format.
When Raw capture is active it will select this format.

Orig-Change-Id: I0f37d3e4341a7768a9c46d38685a4010fea2cc5a
Change-Id: Ie7924b4b063253dd18fc3024137a80618637b50f
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Reviewed-on: http://android.intel.com:8080/128128
Reviewed-by: Gong, Shuguang <shuguang.gong@intel.com>
Tested-by: Gong, Shuguang <shuguang.gong@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>