libcamera2: change stride requirements src from ISP to preview window
BZ: 131368

Until now the source of stride requirements has been the Atom ISP.
After a change in CSS api to accept buffers with diferent strides
we can pass the stride of externally allocated buffers to ISP.

We still need to maintain that all buffers passed to ISP comply
with the ISP stride requirement (i.e. width aligned to multiple of 32).
This will be enforced for external buffers, currently
 - Snapshot buffers allocated in PictureThread
 - Preview buffers allocated in PreviewThread

But this change allows us to pass external buffers with
more stringent stride requirements than ISP, like buffers
going to overlay, to be passed directly to ISP without modification.

Change-Id: I6534ecc032e99b68bed87871e40edb81ba3ccd3d
Signed-off-by: Miguel Verdu <miguel.verdu@intel.com>
Signed-off-by: Juha Alanen <juha.alanen@intel.com>
Reviewed-on: http://android.intel.com:8080/93155
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: Zhu, XiaofeiX <xiaofeix.zhu@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>