libcamera2: temporary setting mechanism
BZ: 133025

This patch introduces a generic temporary setting mechanism which
works similarly as Mutex::Autolock, but for camera settings. In
practice smart pointers and abstraction is used to allow defining
temporary settings which can then be automatically resetted when
the code goes out of scope (e.g. stack instantion) or when sp
ref count drops to zero.

Change-Id: I668f2607c85d5300e99a0d351ae788469e1a4bbb
Signed-off-by: Ukri Niemimuukko <ukri.niemimuukko@intel.com>
Reviewed-on: http://android.intel.com:8080/127699
Reviewed-by: Tuominen, TeemuX <teemux.tuominen@intel.com>
Reviewed-by: Koski, Anttu <anttu.koski@intel.com>
Tested-by: Koski, Anttu <anttu.koski@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>